Migrate LocalizationOptions.swift to the common module
